requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
requests==2.31.0
pydantic>=2.11.0
pytest
pytest-asyncio
httpx
mcp
python-dotenv